+++ Game/RainMeadow.PlayerHooks.cs
@@ -120,11 +120,134 @@ public partial class RainMeadow
         new Hook(typeof(Player).GetProperty(nameof(Player.CanRetrieveSpearFromBack)).GetGetMethod(), DisablePutOnBackWithSpearMasterAbility);
         new Hook(typeof(Player).GetProperty(nameof(Player.CanRetrieveSpearFromBack)).GetGetMethod(), DisablePutOnBackWithSpearMasterAbility);
 
+        IL.Spear.HitSomething += Spear_HitSomething;
+        On.Spear.ChangeMode += Spear_ChangeMode_RemoveDeflectMark;
+        IL.Player.ClassMechanicsArtificer += Player_ClassMechanicsArtificer_OnWeaponParried;
 
         // IL.Player.GrabUpdate += Player_SynchronizeSocialEventDrop;
         // IL.Player.TossObject += Player_SynchronizeSocialEventDrop;
         // IL.Player.ReleaseObject += Player_SynchronizeSocialEventDrop;
     }
+
+    // sync Artificer's parry. Cmon she needs it.
+    public static void PlayArtiParryCustomSound(Weapon weapon) 
+    {
+        weapon.room.PlaySound(SoundID.Spear_Bounce_Off_Creauture_Shell, weapon.firstChunk, false, 1.2f, 2f);
+    }
+    private void Player_ClassMechanicsArtificer_OnWeaponParried(ILContext il)
+    {
+        try
+        {
+            ILCursor cursor = new ILCursor(il);
+
+            cursor.GotoNext(moveType: MoveType.After,
+                i => i.MatchCallvirt<Weapon>(nameof(Weapon.SetRandomSpin))
+            );
+            // There's one occurence for the randomSpin and it's when a weapon is being deflected !
+
+            cursor.Emit(OpCodes.Ldarg_0);
+            cursor.Emit(OpCodes.Ldloc, 11);
+            cursor.Emit(OpCodes.Ldloc, 23);
+            cursor.EmitDelegate((Player player, List<Weapon> weapons, int i) =>
+            {
+                Weapon weapon = weapons[i]; // we grab the weapon from the local variables
+                if (OnlineManager.lobby != null && i == 0) {PlayArtiParryCustomSound(weapon);} // play only in Meadow, not in local
+
+                if (weapon.abstractPhysicalObject.GetOnlineObject() is not OnlinePhysicalObject onlineWeapon
+                    || onlineWeapon.isMine // We don't need to do anything if the spear is ours
+                    || player.abstractCreature.GetOnlineCreature() is not OnlineCreature onlineCreature
+                    || !onlineCreature.isMine) return; // We don't fire if the parry isn't from the client
+                
+                RainMeadow.Debug($"ARTI PARRY !! {onlineCreature}, {onlineWeapon}, {onlineWeapon.owner}, Frame {Mathf.RoundToInt(ARTI_PARRY_MAX_COOLDOWN - player.pyroParryCooldown)}");
+
+                RealizedWeaponState? realizedWeaponState = GetAppropriateWeaponState(onlineWeapon);
+                if (realizedWeaponState is null)
+                {
+                    RainMeadow.Error($"Failed to create the appropriate weapon state for obj {onlineWeapon}");
+                    return;
+                } 
+                
+                onlineWeapon.Lock("parry", onlineWeapon.owner.InvokeRPC(RPCs.Weapon_CreatureDeflect, onlineWeapon, realizedWeaponState, true, i != 0));
+                foreach (OnlinePlayer otherplayer in onlineWeapon.roomSession?.participants ?? [])
+                {
+                    if (otherplayer is not null && otherplayer != onlineWeapon.owner && !otherplayer.isMe)
+                    {
+                        otherplayer.InvokeRPC(RPCs.Weapon_CreatureDeflect, onlineWeapon, realizedWeaponState, true, i != 0);
+                    }
+                }
+            });
+        }
+        catch (Exception e)
+        {
+            RainMeadow.Error(e);
+        }
+    }
+
+    // sync Spear Deflect of creatures, notably by Gourmand, since sometimes creature.SpearStick() is too conditionnal to be 
+    private void Spear_ChangeMode_RemoveDeflectMark(On.Spear.orig_ChangeMode orig, Spear self, Weapon.Mode newMode)
+    {
+        orig(self, newMode);
+        if (newMode != Weapon.Mode.Free
+            && self.abstractPhysicalObject.GetOnlineObject() is OnlinePhysicalObject onlineWeapon
+            && onlineWeapon.IsLocked("deflected"))
+        {
+            onlineWeapon.ClearLock("deflected"); // remove deflect lock when not free anymore, avoiding latent lock junk
+        }
+    }
+    private void Spear_HitSomething(ILContext il)
+    {
+        try
+        {
+            var cursor = new ILCursor(il);
+            cursor.GotoNext(moveType: MoveType.After,
+                i => i.MatchLdarg(0),
+                i => i.MatchCallvirt<Weapon>(nameof(Weapon.SetRandomSpin))
+            );
+            // We go to the very end of the code, where if a weapon has not met the hitting criteria, it'll be deflected.
+
+            cursor.Emit(OpCodes.Ldarg_0);
+            cursor.Emit(OpCodes.Ldarg_1);
+            cursor.EmitDelegate((Weapon weapon, SharedPhysics.CollisionResult result) =>
+            {
+                if (weapon.abstractPhysicalObject.GetOnlineObject() is OnlinePhysicalObject onlineWeapon
+                    && result.obj is Creature target 
+                    && target.abstractCreature.GetOnlineCreature() is OnlineCreature onlineCreature)
+                {
+                    if (!onlineCreature.isMine) 
+                    {
+                        // spear was already deflected, if an RPC comes in, we'll ignore it
+                        onlineWeapon.Lock("deflected"); // Locking manually is abit sketchy, but it's kind of my only way.
+                        RainMeadow.Debug($"{onlineCreature} deflected {onlineWeapon} that isn't mine ! Will ignore the RPC incoming.");
+                    }
+                    else if (!onlineWeapon.isMine) // We don't need to do anything if the spear is ours
+                    {
+                        // We limit that to only the case where the creature is yours but the spear isn't
+                        RealizedWeaponState? realizedWeaponState = GetAppropriateWeaponState(onlineWeapon);
+                        if (realizedWeaponState is null)
+                        {
+                            RainMeadow.Error($"Failed to create the appropriate weapon state for obj {onlineWeapon}");
+                            return;
+                        } 
+                        
+                        onlineWeapon.Lock("parry", onlineWeapon.owner.InvokeRPC(RPCs.Weapon_CreatureDeflect, onlineWeapon, realizedWeaponState, false, false));
+                        
+                        foreach (OnlinePlayer player in onlineWeapon.roomSession?.participants ?? [])
+                        {
+                            if (player is not null && player != onlineWeapon.owner && !player.isMe)
+                            {
+                                player.InvokeRPC(RPCs.Weapon_CreatureDeflect, onlineWeapon, realizedWeaponState, false, false);
+                            }
+                        }
+                    }
+                }
+            });
+        }
+        catch (Exception e)
+        {
+            Logger.LogError(e);
+        }
+    }
+    
     public bool BottomPlayerUsingSpearmasterAbility(Player spearmaster, out Player user)
     {
         user = null!;
@@ -2314,7 +2437,7 @@ public partial class RainMeadow
                 i => i.MatchLdfld<Options>("friendlyFire"),
                 i => i.MatchBrtrue(out _)
                 );
-            c.EmitDelegate(() => (isStoryMode(out var story) && !story.friendlyFire) || (isArenaMode(out var arena) && (arena.countdownInitiatedHoldFire || arena.disableArtiStun)));
+            c.EmitDelegate(() => (isStoryMode(out var story) && !story.friendlyFire) || (isArenaMode(out var arena) && (arena.countdownInitiatedHoldFire || arena.artiStunDistanceMult <= 0)));
             c.Emit(OpCodes.Brtrue, skip);
             c.Index += 6;
             c.MarkLabel(skip);