+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added new gamemode: Drown! Kill and survive to buy your escape. Cooperative or Competitive
 - Scoring update: Empty kills and friendly fire now subtract from the user (in Teams, the killer loses the points). Points are no longer granted to everyone else.
 - Fixed a thing that should happen when you win that hidden thing
+- Fixed winning conditions on that hidden thing
 - Added Import/Export game settings as code with support for FFA, Team Battle, and Drown
 - Slightly improved Void Amoeba's swim speed
 - Added configurable Void Amoeba lethality. Default multiplier is now 4x. You better run.
@@ -13,7 +14,13 @@
 - Added scoring in-game UI for Challenge, Drown, ir if scoring logic is enabled
 - Added Arena Remix tab toggle for disabling Meadow Cosmetics in the lobby
 - Switched the input for banning slugcats from pckup to Shift+Click when using the mouse.
-- Fixed the winning conditions for a secret thing
+- Made Artificer's and Gourmand's parry defender side.
+- Added configurable setting to change Artificer's parry range.
+- Changed configurable setting for Artificer's stun to be a multiplier instead of an on and off checkbox.
+- Added configurable setting to give Artificer's parry some leniency.
+- Added a sound cue to Artificer's parry.
+- Fixed parried spears not being deflected on late parries.
+- Fixed explosive spears still damaging the player when parried.
 ### Modders
 - Added `ExportLocalSettings` and `ImportLocalSettings` virtual functions into `ExternalGameMode` for managing Arena settings 
 ## Story